requirements.txt
Python==3.9
PyTorch==1.10.0
PyG==2.0.3
CUDA==11.3
ogb==1.3.2
tensorboard==2.7.0
rdkit-pypi==2021.9.4
jupyter==1.0
matplotlib==3.4.3
networkx==2.6.3
PyYAML==6.0
scikit_learn==1.0.2
tqdm==4.62.3
h5py==3.11.0
numba==0.56.0
scipy==1.7.0
anndata==0.9.0
numpy==1.22.4
pandas==1.1.5
scanpy==1.7.2